% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimCpp.R
\name{glimCpp}
\alias{glimCpp}
\title{Generalized Local Independence Model (GLIM)}
\usage{
glimCpp(K, N.R, Weights, tol = 1e-07, maxiter = 1e+05, fdb = FALSE)
}
\arguments{
\item{K}{a state-by-problem indicator matrix representing the knowledge
structure. An element is one if the problem is contained in the 
state, and else zero}

\item{N.R}{a (named) vector of absolute frequencies of response patterns.}

\item{Weights}{state-by-problem matrix representing weighting factors to be 
applied to the problem-specific error rates to model their 
dependence on the states.}

\item{tol}{tolerance, stopping criterion for iteration.}

\item{maxiter}{the maximum number of iterations.}

\item{fdb}{feedback during the optimization.}
}
\value{
An object of class \code{glim} having almost the same structure 
        as \code{?blim}
}
\description{
its a generalized local independence model (GLIM) for probabilistic
knowledge structures by maximum likelihood estimation.
}
\details{
See Doignon and Falmagne (1999) for details on the basic local independence
model (BLIM) for probabilistic knowledge structures and 
... for the generalization.
}
\examples{
data(DoignonFalmagne7)
K   <- DoignonFalmagne7$K         # knowledge structure
N.R <- DoignonFalmagne7$N.R       # frequencies of response patterns

# weighting factors
w <- exp(- 0.5 * 
           matrix(c(1, 1, 1, 1, 2, 2, 2, 3, 3, 
                    1, 1, 1, 1, 2, 2, 2, 3, 3, 
                    2, 2, 2, 1, 1, 1, 1, 2, 2, 
                    2, 2, 2, 1, 1, 1, 1, 1, 1, 
                    3, 3, 3, 2, 1, 2, 1, 1, 1) - 1, 9, 5))      

## Fit GLIM 
glimCpp(K, N.R, w)
}
