% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimCpp.R
\name{blimCpp}
\alias{blimCpp}
\title{Basic Local Independence Models (BLIMs) using C++}
\usage{
blimCpp(K, N.R, tol = 1e-07, maxiter = 10000, fdb = FALSE)
}
\arguments{
\item{K}{a state-by-problem indicator matrix representing the knowledge
structure.  An element is one if the problem is contained in the state,
and else zero.}

\item{N.R}{a (named) vector of absolute frequencies of response patterns.}

\item{tol}{tolerance, stopping criterion for iteration.}

\item{maxiter}{the maximum number of iterations.}

\item{fdb}{feedback during the optimization.}
}
\value{
An object of class \code{blim}. See  \code{?blim}.
}
\description{
Fits a basic local independence model (BLIM) for probabilistic 
knowledge structures by maximum likelihood estimation.
}
\details{
See Doignon and Falmagne (1999) for details on the basic local independence
 model (BLIM) for probabilistic knowledge structures.
}
\examples{
data(DoignonFalmagne7)
K   <- DoignonFalmagne7$K         # knowledge structure
N.R <- DoignonFalmagne7$N.R       # frequencies of response patterns

## Fit basic local independence model (BLIM) 
blimCpp(K, N.R)
}
