% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimCpp.R
\name{blimCpp}
\alias{blimCpp}
\title{Basic Local Independence Models (BLIMs) using C++}
\usage{
blimCpp(
  K,
  N.R,
  na.use = c("omit", "missing-as-wrong", "IMBLIM", "MissBLIM"),
  initial_parameter_values = NULL,
  tol = 1e-07,
  maxiter = 10000,
  fdb = FALSE
)
}
\arguments{
\item{K}{a state-by-problem indicator matrix representing the knowledge
structure.  An element is one if the problem is contained in the 
state, and else zero.}

\item{N.R}{a (named) vector of absolute frequencies of response patterns or 
a person-by-problem (item) matrix containing the the individual answers to each item.
An element is one if the problem is solved by the person, and else zero.
Missing values are coded as NA.}

\item{na.use}{a character string indicating a method to deal with missing values (see Details).}

\item{initial_parameter_values}{list containing initial beta, eta and P.K values}

\item{tol}{tolerance, stopping criterion for iteration.}

\item{maxiter}{the maximum number of iterations.}

\item{fdb}{feedback during the optimization.}
}
\value{
An object of class \code{blim}. See  \code{?blim}.
}
\description{
Fits a basic local independence model (BLIM) for probabilistic 
knowledge structures by maximum likelihood estimation.
}
\details{
See Doignon and Falmagne (1999) for details on the basic local independence
 model (BLIM) for probabilistic knowledge structures.
 
 The BLIM cannot handle missing data. In \code{na.use}, however, methods can 
 be specified how missing values are to be taken into account. Besides the 
 trivial methods, \code{"omit"} (data series with missing values are excluded) or 
 \code{"missing-as-wrong"} (missing values are coded as wrong), the missing data can be 
 integrated into the model estimation by specifying the missing data 
 generating process (decimation mechanism). This leads to two extensions of 
 the BLIM, the \code{"IMBLIM"} and \code{"MissBLIM"} (De Chiusole et al. 2015). 
 The \code{"IMBLIM"} (Ignorable Missings BLIM) assumes that the decimation mechanism 
 is independent of the knowledge state K. This is the case, for example, 
 when individuals are presented with only a random subset of items from Q. 
 In the \code{"MissBLIM"} (Non-ignorable missings BLIM), dependencies 
 between the missing values and the knowledge state K are assumed. 
 This means that depending on the mastered items in K, the missing data 
 varies. For example, if one assumes that people who have not mastered an 
 item q are more likely to give no answer than people who have mastered q.
 To capture these dependencies, additional parameters \eqn{\mu = (\mu_q)_{q \in Q}} (\code{mu1})
 and \eqn{\bar{\mu} = (\mu_{\bar{q}})_{q \in Q}} (\code{mu0}) are estimated.
 Here, \eqn{\mu_q} describes the probability that item q, if contained in K, 
 will not be answered, and \eqn{\mu_{\bar{q}}} describes the probability that the 
 item will not be answered if not contained in K.
 Similar to the response patterns themselves, local stochastic independence 
 of the missing values is assumed: It states that for a fixed knowledge 
 state K, not answering or answering one item has no influence on the 
 response behavior for the other items.
}
\examples{
data(DoignonFalmagne7)
K   <- DoignonFalmagne7$K         # knowledge structure
N.R <- DoignonFalmagne7$N.R       # frequencies of response patterns

## Fit basic local independence model (BLIM) 
blimCpp(K, N.R)
}
